Moved soil failure volume to Model namespace
