+++ Editor/AGXUnityEditor/Tools/DeformableTerrainFailureVolumeTool.cs
@@ -1,4 +1,4 @@
-﻿using AGXUnity;
+﻿using AGXUnity.Model;
 using UnityEditor;
 using UnityEngine;
 using GUI = AGXUnity.Utils.GUI;
