--- Editor/CustomEditors/AGXUnity+DeformableTerrainFailureVolumeEditor.cs
@@ -1,13 +1,0 @@
-
-using System;
-using AGXUnity;
-using AGXUnity.Collide;
-using UnityEditor;
-
-namespace AGXUnityEditor.Editors
-{
-  [CustomEditor( typeof( AGXUnity.DeformableTerrainFailureVolume ) )]
-  [CanEditMultipleObjects]
-  public class AGXUnityDeformableTerrainFailureVolumeEditor : InspectorEditor
-  { }
-}
