+++ Editor/CustomEditors/AGXUnity+Model+DeformableTerrainFailureVolumeEditor.cs
@@ -1,0 +1,13 @@
+
+using System;
+using AGXUnity;
+using AGXUnity.Collide;
+using UnityEditor;
+
+namespace AGXUnityEditor.Editors
+{
+  [CustomEditor( typeof( AGXUnity.Model.DeformableTerrainFailureVolume ) )]
+  [CanEditMultipleObjects]
+  public class AGXUnityModelDeformableTerrainFailureVolumeEditor : InspectorEditor
+  { }
+}
